mastodon-ingester/requirements.txt
mastodon.py
kafka-python
